Add memory info to application title
The ComputerInfo library was added to the project in order to gather system memory information. The total amount of physical memory available on the user's machine is now displayed in the application's title in Gigabytes. This new feature provides users with a quick way to understand their system's physical memory capacity.